sniffer/requirements.txt
psutil
scapy
flask
flask-socketio
kafka-python
numpy
tensorflow
scikit-learn
psycopg2
